+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* **2014-10-08**: Make execution of extractors sortable
 * **2014-09-22**: Added the `SeoAwareTrait` to ease creation of SeoAware classes
 
 1.1.0-RC1