+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Changelog
 =========
 
-1.1.0-RC1
+1.1.0-RC2
+---------
 
 * **2014-10-08**: Make execution of extractors sortable
 * **2014-10-08**: Can't use admin extension without burgov/key-value-form-bundle now